:recycle:[refactor]: React.FC 사용 제거
